Limit number of FORMAT items to maximum allowed by bcf1_t n_fmt field
bcf1_t n_fmt is an 8 bit bitfield, so can only store values up to 255.
vcf_parse_format didn't check for overflow of n_fmt while counting
the number of FORMAT items.  This could result in allocation of an
under-sized array, followed by invalid writes when storing the format
information.

To fix this, replace an alloca call with a fixed-sized array (as the
maximum size is small and known).  Remove the code that counts the items
up-front, as it's no longer needed.  Finally, make the loop that iterates
through the FORMAT string bail out with an error if it counts too many
items.