Print message when first creating default cache directory
When $REF_CACHE is unset, if we write reference cache files we write
them within the user's home directory (typically ~/.cache/hts-ref).
Print a message when actually creating this directory, to alert an
unsuspecting user that we may be going to use a lot of disk space
within their home directory.  Hat tip @tk2.

If $REF_CACHE is set, it is non-trivial to determine which level is
a "root" directory to trigger this from.  In this case, we assume
that setting an environment variable indicates that the user is aware
of what's going on.