Avoid random number generator library functions
We are a library, so shouldn't use random()/srandom() as the client
application code may be using them and not expecting someone else to
be interfering with their entropy.

Instead base the unique temporary file name directly on process id,
thread id, current time (using plain C90 functions), and the integer
value of a pointer.