Further updates to multi-threading.
The BGZF writing code now runs in its own dedicated thread.  This
simply permits it to interleave writing time with job dispatching and
other main loop bits.

The writing thread also does periodic hflush, which in turn calls
fdatasync is operating on a local file.  This avoids a much larger
amount to sync at the end when we close the file.

A consequence of the separate write thread is that the thread pool
output queue (aka "results queue") now has a size limit.  Otherwise
failure to write fast enough would build up more and more compressed
blocks in memory.

The net effect of these changes are further speed increases.