Make hts_itr_query find no-coor reads when last reference is unused.
When searching for HTS_IDX_NOCOOR reads, hts_itr_query would return the
wrong result for indexes where the last reference was unused.  Such
references have no entries in the binning index so it it not possible
to return an end offset for them.  Add a loop so that hts_itr_query finds
the last reference with mapped reads.  If found, its ending offset is
returned as the location where the no-coor reads start.