Don't assume order of sequence_ids when finding HTS_IDX_NOCOOR location.
Previous commits b2aab8, 60c22d and cc207d suggest that the order of
sequence_ids in a file may not match the order in the header.  Such
an ordering is permitted by the index format.  In case it happens,
make hts_itr_query check all references when searching for the
HTS_IDX_NOCOOR location to find the one with the highest offset.