Write CRAM ref cache via hFILE rather than stdio
Convert the code to use hFILE, now that hopen() understands "wx".
Mainly motivated by removing paranoid_fclose(): now the only fsync()
call is in hfile.c, so there is only one place we need to provide an
equivalent for Windows, which doesn't have a function of that name.

When aborting, unlink the file *after* closing it, for systems
(i.e., Windows) that can't unlink open files.