Show coordinates in "unsorted positions" error message
Give the user a chance to track down where the problem in their
file is -- and a reason to believe the message is accurate.

TODO Convey the sequence names down to hts_idx_push(); this message
now prints out "...on sequence #2: ...", meaning the second sequence
in the file, which is not unconfusing.