Build DLL and plugins on Cygwin
The Cygwin convention seems to be to just create cyghts-SOVER.dll
without additional symlinks, so we do that.  Linking also creates
an import library, libhts.dll.a, which plays a similar development
role to plain libhts.so; we could make this the true target of the
link rule (cf libhts.so), but currently its name will collide with
the same name used for a plain Windows import library.

NOTE: To use the plugins successfully, there must be only one copy
of libhts.a/cyghts*.dll in the complete executable.  To ensure this,
either make executables use the DLL too (so for htsfile/tabix/etc,
apply s/libhts.a/-L. -lhts/ to the recipe; FIXME) or perhaps we can
link the plugins themselves with some --allow-undefined option rather
than with the import library, as we do on other platforms.

Also ignore plain Windows *.dll for good measure.