First apparent working bgzf_seek implementation.
This also adds new functions to thread_pool, most importantly the
ability to reset a queue.  There is now also a requirement to shutdown
queues before destroying the pool itself.  (You can destroy the pool,
but the queue shutdown will then wedge if it has jobs in flight as
they'll never be processed.)