Fixed bgzf_getc and bgzf_getline to work in multi-threaded mode.
This means in theory samtools faidx and htslib's tabix could be
multi-threaded.  In practice the gain is slight as both of these
commands spend a lot of cpu doing more than decompressing.  I could
get 60% gain to faidx rate and ~100% gain to tabix index rate.

Maybe worth keeping, but currently not enabled.  The main change is in
bug fixing the functions.