Added support for multi-threaded BAM indexing.
Also enabled multi-threaded decoding of text-based bgzf formats (SAM,
VCF, maybe others?).  Is there an easy way to detect when the voidp
has been utilised in htsfile and if so whether it is to a bgzf rather
than some other generic type?  I'm not particularly happy with the
sam.gz MT implementation.