PTHREAD_MUTEX_RECURSIVE_NP vs PTHREAD_MUTEX_RECURSIVE.
The NP is the documented type and stands for non-portable.  As
expected, it's not portable!  It doesn't work on MacOS X, but removing
the _NP works on more systems.

TODO: Figure out how to avoid getting in the situation of locking
multiple times.  It comes about through pool->queue and queue->pool
interactions, both of them sharing the same mutex.