Generate config.h.in with autoheader
Previously we hand-edited config.h.in after using autoheader, to avoid
the dross -- especially PACKAGE_VERSION, as we compute that ourselves
(via git describe) in the Makefile.  However that's used only within the
Makefile, so defining it for C code via config.h is not a major problem.
(If it becomes one, we can add AH_BOTTOM([#include "config_undefines.h"])
or so, or not reuse the conventional PACKAGE_VERSION macro name.)