Add configure check for fdatasync()
Enables the existing HAVE_FDATASYNC conditional in hfile.c.

OS X has a library symbol named fdatasync() but no declaration in
the headers, and it is unclear whether its function has the expected
fdatasync(2) signature.  So we additionally use AC_CHECK_DECL to
avoid detecting this symbol on OS X.