Renamed various thread pool structs/functions.
t_pool        -> hts_tpool
t_pool_result -> hts_tpool_result
t_pool_queue  -> hts_tpool_process

The way we think about queue is that in "foo |gzip| bar" and
"bar |gunzip| foo", the "gzip" and "gunzip" are process names and the
"|" symbols in "|gzip|" are the input and output queues associated
with this process.  Our old queue was actually a pair of queues (the
pipe symbols) with the entire struct modelling the I/O to the process.

Technically it's not the process itself but rather a wrapper around
the process, as that is another C func (whatever was given during
a dispatch call), but in lieu of any better term this will do.
(Please offer other ideas if you have them.)

Also added hts_tpool_kill so that hts_tpool_destroy is now simpler in
operation, with only one argument required.