Added callback + client data hooks to pileup iterators and pileup struct.
The mechanism here is to permit a pileup user to locally cache some
data when a new bam record is in flight, attached to the pileup struct
passed to the caller, and to tidy it up once it goes out of scope.
This greatly simplifies some programming methods and also permits more
efficient code in some places where doing the right thing (caching
values) was just too hard to make it worth while.

Also added an mplp version of the callbacks too, which simply iterates
over the plp ones.