Allow plugins to select RTLD_LOCAL or RTLD_GLOBAL
If the plugin does not provide the requested symbol, fall back to
reopening it with RTLD_GLOBAL and searching for a uniquified symbol,
<symbol>_<filename_basename>.

Most plugins should work with RTLD_LOCAL, but occasionally RTLD_GLOBAL
is needed.  For example, with iRODS 4.1.x, iRODS itself has plugins
(libtcp.so et al) that need to resolve iRODS symbols linked into
whatever is invoking it, i.e., HTSlib's hFILE plugin.