Protect against sequences starting beyond reference end.
This was previously causing the slice reference md5sum to crash.

Fixes samtools/samtools#600, but note this perhaps isn't a perfect
fix.  The test data for that issue is broken, but ideally we should be
able to reproduce the broken input after round-tripping.  Currently we
lose track of MD/NM tags.

However making cram totally lossless even when faced with invalid data
is an issue for another date.