Embed version number directly in hfile_libcurl plugin
When built as a separate plugin, it is incorrect to use hts_version()
to report HTSlib's version number, as we could be used with a variety
of HTSlib versions.  Instead embed our own copy of HTS_VERSION as of
when we were compiled.  (But just use hts_version() when plugins are
disabled and hfile_libcurl.o is built within HTSlib.)

Embed an SCCS-style @(#) version string to faciliate examination via
strings(1) | grep '@(#)' or what(1).