Discard distant chunks based on binning index, not linear index
Contrary to the description in SAMv1.pdf:

    In the linear index, for each tiling 16384bp window on the
    reference, we record the smallest file offset of the alignments
    that start in the window.

it seems the linear index actually records the smallest file offset of
an alignment that *overlaps* the window.  This could start at a position
before the bin-to-the-right we're looking at, and lead to clipping chunks
containing reads in the desired region.

Instead take max_off from a chunk fully within the bin-to-the-right;
due to sorting the first one will be the tightest bound, and as the list
of chunks is set directly, we don't need to check for 0.  Fixes #405.