Handle VCF lines with misssing `FORMAT=.`
For example, this is a valid VCF line

```
1	300	.	C	A	.	PASS	.	.	.	.
```

Previously this would emit a warning saying:
`[W::vcf_parse_format] FORMAT '.' is not defined in the header, assuming Type=String`
and internally we would have a new `FORMAT=.` tag.
This will now be recognised as missing.

htslib already writes out such lines when  `n_fmt == 0` and `n_samples > 0`

Mixing missing and non-missing FORMAT tags (e.g. `.:GT` or `GT:.:AD`) is not allowed.

See conversation in #409