Fix hFILE write-after-read bug
The "desired position is within our read buffer" hseek() optimisation
also breaks hread()-hseek()-hwrite() on files opened for update, as
hwrite() expects the hseek() to have left an empty direction-agnostic
buffer.  So we disable the optimisation on files opened for update.

Alternative fixes would be to have hflush() discard any read buffer and
do a backend->seek(begin-end, SEEK_CUR) (or rather the equivalent SEEK_SET)
to correct the backend file position, and require callers to use hflush()
when switching between reading and writing in either direction; or (hat
tip @daviesrob) have hputc()/hputs()/hwrite() check for a non-empty read
buffer, and discard it and do a corrective seek if necessary.