Make hts_expand handle realloc failure a bit better.
The hts_expand() and hts_expand0() macros don't return a value, and will
set (ptr) to NULL if realloc fails.  This could lead to segmentation faults
if callers don't check the value of (ptr).  As few if any do, and some
callers are in external packages, the best solution is to make the
hts_expand macros print an error message and call exit(1).

It's not ideal behaviour for library code, but fixing this in any other
way is just too hard.

This implementation removes the assumption that (m) and (n) are of type
int.  It should work with any integer type no bigger than size_t.