Add support for Google Cloud Storage pseudo-URLs
Rewrite gs: pseudo-URLs to http/https URLs, adding an Authorization
header for Google Cloud Storage.

At present, the access token is simply taken from a bespoke environment
variable, $GCS_OAUTH_TOKEN, as per PR #390.  (TODO) This should be
replaced with pulling the information from shared configuration files
or obtaining it via other authentication infrastructure.

In configure.ac, this is before AC_SYS_LARGEFILE to keep the options
in ./configure --help in alphabetical order.

Passing the va_list to hopen(..., "va_list", ...) in gcs_vopen() is
painful as passing a va_list (as opposed to a va_list*) via .../va_arg
proved impossible, and we need a local va_list object rather than a
parameter with a possibly-decayed type in order to form a va_list pointer.
See the following for details:
http://stackoverflow.com/questions/4958384/what-is-the-format-of-the-x86-64-va-list-structure
http://stackoverflow.com/questions/8047362/is-gcc-mishandling-a-pointer-to-a-va-list-passed-to-a-function