Change bam1_core_t::n_cigar from uint16_t to uint32_t
As various CIGAR-related loops and API functions use int, the practical
limit without further changes is 2^31-1 rather than 2^32-1.  As bam_init1()
uses calloc(), the unused field is initialised to 0 and we don't need to
bother re-zeroing it in sam_read1() et al.

Until the BAM format is extended to deal with it (samtools/hts-specs#40),
trying to write >64K CIGAR operations to a BAM file remains an error.

Implements and fixes #437.  (Update htslib/sam.h copyright notice for
changes in 2015 and 2016.  MinGW does not have EOVERFLOW.)