Prevent reads past the end of the VCF header.
Modify bcf_hdr_read to ensure the header is always NUL-terminated.

Fix returned length in bcf_hdr_parse_line when it reaches a NUL character.
Previously it would include the NUL for lines which did not finish with \n.
This could cause a later call to the same function to read beyond the end
of the input buffer.