Add function bcf_record_check() to validate bcf records
Check contig and tag ids are within the valid range given in the header.

Check for sensible data types.

Ensure decoding doesn't run off the end of the shared and indiv strings.

It's currently only called by bcf_read(), but might want to be used in
other places (or possibly even exposed in the API).