Alter bam1_t data layout so that CIGAR data is 32-bit aligned
Add extra padding NULs after qname so that the CIGAR data in memory is
aligned on a 32 bit boundary.  These NULs are included in l_qname so
that accessor macros are unchanged (as is code that accesses this data
without using the macros, sigh), and a new l_extranul field is added
that counts the extra NULs beyond the existing terminator NUL.

Add tests that read and write SAM, BAM, and CRAM.  Check sizeof(bam1_t),
taking care to allow for a possible extra 4 bytes of alignment padding
(e.g., on 64-bit platforms).

Fixes #400.