Permit CRAM lossy_names mode to accept TLEN 0 or TLEN +/- 1.
The lack of TLEN value (eg "* 0 0" for the 3 SAM fields) or TLEN being
out by 1 causes the CRAM encoder to encode reads in "detached" state
so it can store TLEN verbatim.  This process broke the lossy name
encoding, so we're no longer quite so precise in our round-trip for
TLEN if we're asking for lossy read names.

Ideally these TLEN options would be a separate and orthogonal option
to the lossy_names option, but for now they are tied together for
simplicity.