Ensure max_off is -1 when end bin overflows
When the end coordinate is larger than the index's maximum coordinate,
the starting bin for computing max_off is somewhere on the row beyond
the index's bottom row.  Disable max_off rather than using a random
parent bin of the bin somewhere below the bottom row.  (The min_off
computation has a similar problem, but for min_off it's immaterial.)

Fixes #455.