Add hts_endian.h to convert little-endian bytes to/from native integers.
Also handles unaligned access.  This can be controlled by setting
HTS_ALLOW_UNALIGNED - if 0, unaligned access is disabled; if 1 it is allowed.
If unset, it will be enabled on intel x86-like platforms.

cram/cram_encode.c and cram/os.h are modified to use HTS_ALLOW_UNALIGNED
instead of ALLOW_UAC (which could not be disabled).

One of the macros HTS_LITTLE_ENDIAN or HTS_BIG_ENDIAN is defined on
platforms known to be little- or big- endian respectively.  HTS_ENDIAN_NEUTRAL
can be defined to disable this.  Code in hts_endian.h is intended to
be endian-neutral by default, so it should still work even where
endian-ness detection fails.

Includes unit tests and documentation.