Fixed dead-lock case in seek + multi-threaded decode.
If the bgzf_mt_reader() function was blocked (queue input is full)
dispatching the very last block of the file, and the main thread then
sends a SEEK command, the reader is woken up but does not process the
seek request.  Instead it wedges on the next dispatch, marking EOF,
instead.

The reversal of the blocks of code appears valid given we have one
read-ahead thread per file open each with its own queue of blocks to
decode.  Hence it should not be possible for the very first dispatch
call to block.

This hopefully resolves the issue #537, subject to further testing by
the submitter.