More thread pool fixes (see issue #537).
These come with some test code for thrashing the bgzf multi-threading,
but the tests are not normal tests and do not get executed by make
check.  Instead they are designed to be run manually to look for
lockups and to be executed under valgrind (memcheck, drd, helgrind
tools) to look for issues.

The bugs fixed include race conditions on both reader and writer
between shutting down the thread pool and starting the reader/writer
itself and races between shutting down and the reader/writer exiting.

Also fixed bgzf_seek when given a raw gzip stream.  It now returns
with BGZF_ERR_MISUSE error as we cannot seek into those.