Report missing BGZF EOF blocks
Make bgzf_read_block report missing EOF blocks while reading in
single-threaded mode. The multi-threaded code is updated so that it uses
the same field in struct BGZF to record having seen an EOF block.  We
also set a bit when we discover the EOF block is missing so as to
avoid repeated warnings on the subject.

The bizarre 2-bit booleans in struct BGZF are modified to be 1 bit long
so we can repurpose some of the bits.  We can get away with this as we
are changing the ABI.