Add bgzf_index_load_hfile and bgzf_index_dump_hfile
These read and write .gzi indexes in the same way as bgzf_index_load and
bgzf_index_dump, but allow writing to an existing file handle.

bgzf_index_load and bgzf_index_dump are modified to simply open an hFILE
and call the new functions.

Doxygen documentation is added for the new functions, and improved for the
existing ones.