Remove abort on corrupt aux data, pass errors up instead
Remove abort from skip_aux() when it finds an aux record of unknown type.
Also add some tests for records that are longer then the space available
for them.  Make it return NULL if anything is wrong.

Make bam_aux_get() and bam_aux_del() report an error if the aux data is
founbd to be corrupt.  bam_aux_get() now sets errno so callers can tell
the difference between broken aux data and tags that are not present.
errno is used so that the API for bam_aux_get() is essentially unchanged -
callers that do not check will still work, albeit while silently ignoring
the problem.