Make sam_format1() fail it it finds an invalid aux type
Add a missing catch for invalid aux types.

If hts_verbose is set it will now print a message if it fails due to
invalid aux data.

Also fix a FIXME - check if there is enough aux data left the correct way.

Thanks to Chris Saunders at Illumina for reporting problems with
sam_format1() and skip_aux() on broken records.