Fix endianness, integer type and memory safety issues in index metadata
Index metadata stored in an hts_idx_t struct was byte-swapped for
tbi indicies but not for csi.  Make them both the same by leaving the
data in little-endian order, and make tbx_index_load2() access the
data in an endian-neutral manner.

Functions hts_idx_get_meta() and hts_idx_set_meta() are changed to
treat the metadata length as an unsigned value (uint32_t) to match
the integer type that is used to store the length in the hts_idx_t
struct.  hts_idx_set_meta() is also changed to return an int
so callers can detect if it ran out of memory.

hts_idx_load_local() and hts_idx_set_meta() now ensure that the stored
meta-data is always followed by a NUL.  This is to prevent tbx_index_load2()
from running off the end of the metadata if there is an unterminated
string in the list of sequence names.

get_tid() is made to return -1 if it fails to add a hash entry.  A few
other calls to malloc are also made safer.