Remove explicit -lz -lm link flags; add to LIBS instead
Add -lz and -lm to $(LIBS) via $(htslib_default_libs).  This means they
don't need to be explicitly need to be on link lines.  LIBS will be updated
by config.mk if configure is used.  configure.ac is updated to ensure
configure supplies -lz and -lm.

MacOS libhts.dylib may now get -lm on its link line, which is not strictly
needed, but harmless (libm.dylib is a symbolic link to libSystem.dylib).