Fixed a bug in bcf_fmt_array with garbage allele frequencies. The special value 0x7f800001 (NaN), used to indicate a missing value, was changed to 0x7fc00001 (also NaN) when the function le_to_float returned.
This problem might be specific to MSVC and the calling convention where the return value is passed on the FPU register stack. In that process the value seems to be converted from signaling NaN to quiet NaN.

The fix is to keep the value as uint32 while testing for special values.

Reported as a bug in samtools:
https://github.com/samtools/hts-specs/issues/145