Bug fix in VCF/BCF header parsing
When VCF header contained a line that could not be parsed, all samples
were discarded. For example, in a well-formed VCF all header lines must
be of the form ##key=value and if they are not, the parsing would stop
prematurely and consequently all genotype fields dropped:

    ##fileformat=VCFv4.2
    ##FORMAT=<ID=GT,Number=1,Type=String,Description="Genotype">
    ##contig=<ID=1>
    ##Incorrect comment line
    #CHROM POS ID REF ALT QUAL FILTER INFO FORMAT Sample
    1      1   .  G   T   .    .      .    GT     0/1

This commit skip malformed header lines, printing a warning.

On non-recoverable errors, bcf_hdr_parse now returns negative
value which is now checked by all callers.