Ensure bcf_hdr_parse_line() always finishes at the next newline or NUL
Terminate the search for an '=' character if a newline is found.  This
stops bcf_hdr_parse_line() from incorrectly running into the next line
when '=' is missing.

On structured header lines, drop all characters between the closing '>'
and the next '\n' or NUL, not just spaces.  This ensures *len is set to
the correct line length, and the next call starts parsing at the
correct location.  If any non-whitespace characters are seen, a warning
is printed.

(Minor) Remove the need for an intermediate buffer when printing the
'Could not parse the header line' error message.