Fix length calculation in bam_read1 and possible memory leak
Add missing c->l_extranul when checking how much space is needed for
the name, cigar, seq and qual data.  Fixes #547 where the block length
in the bam record was short by a few bytes but the check failed because
the extra NULs after the name were not being counted.  This could cause
a buffer over-run by up to three bytes (but only when l_data was close
to a power of two).

The new version does not use bam_get_aux() to avoid some dubious
pointer arithmetic (especially when d->data has not been allocated).

Also ensure that memory is not leaked and b->m_data is left unchanged
if the realloc on b->data fails for some reason.