Experimental: remove check for HAVE_FSEEKO.
Does this fix AppVeyor builds?  Regardless it's wrong to check this
macro now we've removed it again from configure.ac, but I don't (yet)
know if we need to remove it here too or whether we need to keep this
as-is and add it back to configure.ac again.  Hopefully the former,
especially given we have no support for MSVC anyway.