Update multipart to work with revised htsget specification
PR samtools/hts-specs#230 updates the htsget specification so
that all htsget JSON starts with a protocol identifier.  This
is implemented by wrapping the payload inside a JSON object
with the single key "htsget".  Update multipart.c and
hts_detect_format() in hts.c to understand the revised format.

As the protocol now has an official name, change some mentions of
'json' to 'htsget'.  One of these is in the htsExactFormat
enumerated type, which is in htslib/hts.h.  This is left in
place for API compatibility, but marked as deprecated (if using
a compiler that supports this).