Added BGZF CRC32 checking.  Fixes #402.
Note this normally is an insignificant overhead, but if decoding
uncompressed BAMs it becomes dominant.  Eg perf on samtools flagstat
shows:

 70.47%  samtools.tmp  libz.so.1.2.8       [.] crc32
  8.91%  samtools.tmp  [kernel.kallsyms]   [k] 0xffffffff8104f45a
  8.40%  samtools.tmp  libc-2.19.so        [.] __memcpy_sse2_unaligned
  4.80%  samtools.tmp  samtools.tmp        [.] bgzf_read
  3.95%  samtools.tmp  samtools.tmp        [.] bam_read1

One solution is either to have an option to not check (I don't like
this), or to replace the default zlib crc algorithm with a much more
performant alternative.  See #467 for discussion.  Libdeflate's SSE
crc is around 11x faster, meaning this overhead would be around 20%
instead.

Perhaps it's time we had a build option linking against libdeflate
anyway as it's almost certainly better for all of the work instead of
just this bit.
https://encode.ru/threads/2482-libdeflate-a-new-optimized-DEFLATE-implementation