Added libdeflate as an option.
Currently there is no autoconf setup.  Compile with:

    make CPPFLAGS="-DHAVE_LIBDEFLATE -I$HOME/ftp/compression/libdeflate" LDFLAGS="-L$HOME/ftp/compression/libdeflate" LIBS="-ldeflate"

This overly complex and ideally should just be a single parameter:

    ./configure --with-libdeflate=$HOME/ftp/compression/libdeflate
    make

It fails one test in htslib; rebgzip, which is fundamentally broken by
design and cannot be fixed.  I'm tempted to suggest we just skip the
test in this scenario.

It fails several in samtools, but these are probably also broken tests
which would also fail if we switched to e.g. Intel or CloudFlare
zlibs.

Speed tests on a small file:

u.BAM -> BAM    Was 13.5s, now 6.3s.
BAM -> u.BAM    Was  2.5s, now 1.8s.