Make hts_test_and_fetch check for fwrite() errors
The interface is improved so that it's possible to tell the
difference between failing to open the remote files and other
errors.  This stops hts_idx_getfn() from attempting a retry
which is not going to work.

hts_idx_getfn() will no longer attempt another fetch if it failed
to find a '.' in the input file name.

Fixes issue #580