Fix cram integer overflows
Use comparisons instead of subtraction in code_sort().  The result
of the subtraction may be too big for a signed integer.

Reject code lengths > 31 bits in cram_huffman_decode_init().
Anything longer than this causes an overflow in the value used to
store the code.

Ensure that huffman code values are all inside the specified bit
length.  It's possible to overflow this by having too many
symbols share the same code length.

Make `len` an int64_t in cram_read_SAM_hdr() so it cannot overflow.

Ensure that compressed and uncompressed sizes are the same for RAW
in encodings in cram_read_block().  The specification says they
should be, and the code relies on it in places.