Add Bearer token support to hfile_libcurl (mainly for htsget)
Make hfile_libcurl check environment variable HTS_AUTH_LOCATION
to get the location of a file where an OAUTH2 style Bearer token
can be found.  If the location includes the string '%h' then
it is replaced by the host name (and port, if present) part
of the URL being accessed.  This allows tokens for different
destinations to be stored separately.

The token can take one of two forms - a single line containing
the access token itself, or an access token response as described
in RFC 6750 section 4 which has the form of a JSON object.
In the second case, the value of the "expires_in" field will be
stored, and an attempt will be made to re-read the token if it
has become stale (the assumption being that the token comes
from some sort of agent that can refresh it).

The token file is opened using hopen(), so as well as a plain
file it could also be an http(s):// url that points to an agent
that manages token refresh and provides them to clients via HTTP.
This hopen() is prevented from reading an access token itself,
otherwise HTSlib would get into an infinite recursion.  This is
done by adding 'R' to the mode string.  The method was chosen
because it will simply be ignored by any hfile handlers that
don't understand it, and it will be passed on by handlers that
invoke hfile_libcurl indirectly (e.g. hfile_s3, hfile_gcs).

To prevent token disclosure, Authorization headers will only be
added for https URLs.  This can be overridden by setting an
environment variable (see code for details).