Fix data: urls in multipart requests
Not all URLs accessed by multipart.c go through hfile_libcurl.
As we want to pass an hfile_libcurl-specific hopen() parameter
and there is currently no way to find if a parameter will be
accepted, we check to see if the URL will be handled by
hfile_libcurl before calling hopen().

find_scheme_handler() is made non-static, but not made part of
the public API.