Add BAM version 2 support
BAM version 2 increases the size of n_cigar to 32 bits.  It
adds `bam2_hdr_flags` and `bam2_flags` (on the alignment records).
These are currently reserved for future extensions so are
written as zero and checked to ensure they are zero on reading.

The header magic number is changed to "BAM\2".

To select version 2 when writing, it's necessary to either call
`hts_set_opt(fp, HTS_OPT_VERSION, "2.0")` before writing the BAM
header, or add a 'BV:2' tag to the SAM `@HD` line.  On reading, the
version is set automatically depending on the file magic number.

test.pl is updated to test SAM -> BAM2 -> SAM round trips.