Add configure options to enable warnings and optionally Werror (#632)
Add option to enable warnings (on by default) and -Werror or equivalent
(off by default).  Original macros to do this came from ac-archive
on SourceForge:
 http://ac-archive.sourceforge.net/ac-archive/vl_prog_cc_warnings.html
via libstaden_read but have now been heavily modified.

Werror only works on GCC compatible and Sub compilers at the moment.
The -Werror or equivalent flag is only added to CFLAGS at the end
of the configuration process, as it can interfere with some tests.