Fixes to PR#590.
The buffer used in preload is now freed.  If we want to take ownership
of it and avoid it being freed, the new hfile_mem_steal_buffer
function can be called.

Also changed the prototype of hfile_mem_get_buffer to return the
buffer directly instead of via a pointer to a pointer as it's
simpler and permits inline usage.

Changed preload to work on more than straight hFILE_fd, although in
the current incarnation it still won't work on anything that attempts
to use the varags hopen interface for anything else.