Avoid unnecessary delays in wait_perform()
Always call curl_multi_timeout(), even if curl_multi_fdset()
returned a negative value for maxfd.  It frequently returns
zero in this case, indicating that work still needs to be done.
The timeout is also set to a maximum of 100ms when maxfd < 0, as
recommended by the curl_multi_fdset() manual page.

If fp->perform_again is true, the calls to curl_multi_fdset() and
curl_multi_timeout() are skipped completely as the values
calculated were not used.

Fixes #633 (Curl vs knet speeds).