Delay most seek operations in hfile_libcurl
"Seek" on http(s) connections is a lot of work as it involves
making a new request and throwing away the existing one.  This
work is wasted if two or more seeks happen without any reads. An
example of where this can happen is the BGZF block cache, which
calls hseek() from load_block_from_cache() to adjust the file
pointer to the end of the returned block.

To avoid this wasted effort, make libcurl_seek() perform the
full process the first time it is called on each file handle (to
test that seeking actually works), then for all following simply
record the location where the next read should start and do the
actual seek on the next call to libcurl_read().  Testing on the
first call means non-seekable streams should cause hseek() to
return an error as would normally be expected.

The common code for restarting the connection is now in the
function restart_from_position().  The error handling is improved
so that it should always return a suitable value in errno when
called from libcurl_read().