Fixed return value of cram_get_bam_seq and fixed a bug in required_fields option.
1. bam_construct_seq returned 0 instead of number of bytes written to
the bam_seq_t struct.  This in turn bubbled up to cram_get_seq
returning only the number of auxiliary bytes written to instead of the
total record size.

Under specific circumstances (no RG tag, no NM/MD requested, and no
other aux fields), this meant cram_get_bam_seq returned 0 on a
successful read.  This in turn was interpreted as EOF by the iterator.
Given bam_read1 returns the size of the bam structure it feels
appropriate for the cram version to mirror it, as originally intended.

2. If we use the required_fields option to exclude sequence data, then
we also exclude loading of the reference (which is a useful
optimisation).  However this then also meant the cram_decode_seq
function wasn't correctly modifying the alignment end, leaving it set
to alignment start.

The consequence is range queries combined with required_fields usage
could miss reads which start before the range and end within it.

Fixes #640